some file renames, minor refactoring
