+++ src/management-system-v2/app/shared-viewer/document-content.module.scss
