+++ src/management-system-v2/components/doc-element-sections.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Image, Table, Typography } from 'antd';
-import styles from '@/app/shared-viewer/process-document.module.scss';
+import styles from '@/app/shared-viewer/document-content.module.scss';
 import { ElementInfo } from '@/app/shared-viewer/table-of-content';
 
 const { Title } = Typography;
