+++ src/management-system-v2/components/instance-doc-tables/final-variables-table.tsx
